refactor: address code review
